# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice para gestão de cobranças (inicialmente integração Asaas; no futuro bancos diversos).",
        "title": "Charges Service",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
...
        "/v1/asaas/charges/{id}/digitable-line": {
            "get": {
                "description": "Retorna a linha digitável (identificationField) para uma cobrança no Asaas.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Linha digitável do boleto (Asaas)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da cobrança no Asaas (payment id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/asaas/charges/{id}/pix-qrcode": {
            "get": {
                "description": "Retorna o QRCode Pix (encodedImage/payload) para uma cobrança no Asaas.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "QRCode Pix (Asaas)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da cobrança no Asaas (payment id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
}